+++ src/cosmosdb-preview/azext_cosmosdb_preview/custom.py
@@ -779,7 +779,8 @@ def cli_cosmosdb_create(cmd,
                         enable_materialized_views=None,
                         enable_burst_capacity=None,
                         enable_priority_based_execution=None,
-                        default_priority_level=None):
+                        default_priority_level=None,
+                        enable_per_region_per_partition_autoscale=None):
     """Create a new Azure Cosmos DB database account."""
 
     from azure.cli.core.commands.client_factory import get_mgmt_service_client
@@ -835,7 +836,8 @@ def cli_cosmosdb_create(cmd,
                                     enable_materialized_views=enable_materialized_views,
                                     enable_burst_capacity=enable_burst_capacity,
                                     enable_priority_based_execution=enable_priority_based_execution,
-                                    default_priority_level=default_priority_level)
+                                    default_priority_level=default_priority_level,
+                                    enable_per_region_per_partition_autoscale=enable_per_region_per_partition_autoscale)
 
 
 # pylint: disable=too-many-branches
@@ -869,7 +871,8 @@ def cli_cosmosdb_update(client,
                         enable_materialized_views=None,
                         enable_burst_capacity=None,
                         enable_priority_based_execution=None,
-                        default_priority_level=None):
+                        default_priority_level=None,
+                        enable_per_region_per_partition_autoscale=None):
     """Update an existing Azure Cosmos DB database account. """
     existing = client.get(resource_group_name, account_name)
 
@@ -962,7 +965,8 @@ def cli_cosmosdb_update(client,
         enable_materialized_views=enable_materialized_views,
         enable_burst_capacity=enable_burst_capacity,
         enable_priority_based_execution=enable_priority_based_execution,
-        default_priority_level=default_priority_level)
+        default_priority_level=default_priority_level,
+        enable_per_region_per_partition_autoscale=enable_per_region_per_partition_autoscale)
 
     async_docdb_update = client.begin_update(resource_group_name, account_name, params)
     docdb_account = async_docdb_update.result()
@@ -1167,7 +1171,8 @@ def _create_database_account(client,
                              enable_burst_capacity=None,
                              source_backup_location=None,
                              enable_priority_based_execution=None,
-                             default_priority_level=None):
+                             default_priority_level=None,
+                             enable_per_region_per_partition_autoscale=None):
 
     consistency_policy = None
     if default_consistency_level is not None:
@@ -1306,7 +1311,8 @@ def _create_database_account(client,
         enable_materialized_views=enable_materialized_views,
         enable_burst_capacity=enable_burst_capacity,
         enable_priority_based_execution=enable_priority_based_execution,
-        default_priority_level=default_priority_level
+        default_priority_level=default_priority_level,
+        enable_per_region_per_partition_autoscale=enable_per_region_per_partition_autoscale
     )
 
     async_docdb_create = client.begin_create_or_update(resource_group_name, account_name, params)