+++ src/cosmosdb-preview/HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ===============
 0.27.0
-
+* Add support for Per-Region Per-Partition Autoscale. '--enable-per-region-per-partition-autoscale' parameter can be used during account create/update.
 ++++++
 0.26.0
 * Add '--enable-priority-based-execution' and '--default-priority-level' parameter for create/update database account.