+++ src/cosmosdb-preview/HISTORY.rst
@@ -4,6 +4,8 @@ Release History
 ===============
 0.27.0
 * Add support for Per-Region Per-Partition Autoscale. '--enable-per-region-per-partition-autoscale' parameter can be used during account create/update.
+* Add support for online mode in container copy job.
+
 ++++++
 0.26.0
 * Add '--enable-priority-based-execution' and '--default-priority-level' parameter for create/update database account.