+++ src/cosmosdb-preview/HISTORY.rst
@@ -6,6 +6,9 @@ Release History
 * Add support for Per-Region Per-Partition Autoscale. '--enable-prpp-autoscale' parameter can be used during account create/update.
 * Add support for Restore with Time-To-Live Disabled. '--disable-ttl' parameter can be used during restore.
 * Add support for online mode in container copy job.
+* Add support for minimum allowed TLS version configuration
+* BREAKING CHANGE: az cosmosdb create/update: Rename --enable-public-network true/false to --public-network-access ENABLED/DISABLED/SECUREDBYPERIMETER
+* BREAKING CHANGE: az cosmosdb restore: Rename --enable-public-network true/false to --public-network-access ENABLED/DISABLED
 
 ++++++
 0.26.0